backend/app/schemas/project.py
from typing import Optional
from pydantic import BaseModel
from datetime import datetime


class ProjectBase(BaseModel):
    name: str
    description: Optional[str] = None
    color: str = "#3B82F6"
    is_active: bool = True


class ProjectCreate(ProjectBase):
    pass


class ProjectUpdate(ProjectBase):
    name: Optional[str] = None
    is_active: Optional[bool] = None


class ProjectInDBBase(ProjectBase):
    id: int
    user_id: int
    created_at: datetime

    class Config:
        from_attributes = True


class Project(ProjectInDBBase):
    pass


class ProjectInDB(ProjectInDBBase):
    pass